DDF-4451 Refactor imports/exports
